Move time formatting into module
