+++ lib/rspec/matchers/built_in/base_matcher.rb
@@ -113,6 +113,33 @@ module RSpec
           alias present_ivars instance_variables
         end
 
+        module TimeFormatting
+          TIME_FORMAT = "%Y-%m-%d %H:%M:%S"
+
+          if Time.method_defined?(:nsec)
+            def format_time(time)
+              time.strftime("#{TIME_FORMAT}.#{"%09d" % time.nsec} %z")
+            end
+          else # for 1.8.7
+            def format_time(time)
+              time.strftime("#{TIME_FORMAT}.#{"%06d" % time.usec} %z")
+            end
+          end
+
+          DATE_TIME_FORMAT = "%a, %d %b %Y %H:%M:%S.%N %z"
+          # ActiveSupport sometimes overrides inspect. If `ActiveSupport` is
+          # defined use a custom format string that includes more time precision.
+          def format_date_time(date_time)
+            if defined?(ActiveSupport)
+              date_time.strftime(DATE_TIME_FORMAT)
+            else
+              date_time.inspect
+            end
+          end
+        end
+
+        include TimeFormatting
+
         # @private
         module HashFormatting
           # `{ :a => 5, :b => 2 }.inspect` produces:
